requirements.txt
numpy~=1.26.4
pillow~=10.2.0
Flask~=3.0.0
scikit-learn~=1.5.1